backend/mage/requirements.txt
# Core dependencies
requests>=2.28.0
mage-ai>=0.9.0

# Cloud storage
google-cloud-storage>=2.0.0

# XML parsing
lxml>=4.9.0  # Optional: faster XML parsing

# Geospatial and data processing - essential for wetlands_merge_grid.py
# These dependencies are also included in the Dockerfile
# If you encounter import errors, uncomment the relevant lines
shapely>=2.0.0
geopandas>=0.9.0
dask[dataframe]>=2023.0.0
dask-geopandas>=0.1.0
pyarrow>=10.0.0
distributed>=2023.0.0
pyogrio>=0.6.0
libpysal>=4.0.0
psutil>=5.9.0  # System monitoring

# Add other dependencies as needed
# package1>=1.0.0
# package2>=2.0.0
